Separating template implementations from CUDA utility header.
