+++ include/lbann/utils/CMakeLists.txt
@@ -23,5 +23,8 @@ set_full_path(THIS_DIR_HEADERS
   timer.hpp
   )
 
+# Add the subdirectories
+add_subdirectory(impl)
+
 # Propagate the files up the tree
 set(HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}" PARENT_SCOPE)
