+++ include/lbann/utils/cuda.hpp
@@ -34,9 +34,6 @@
 #include <cuda.h>
 #include <thrust/memory.h>
 #include <thrust/detail/allocator/tagged_allocator.h>
-#ifdef __CUDACC__
-#include <cuda_fp16.hpp>
-#endif // __CUDACC__
 
 // -------------------------------------------------------------
 // Error utility macros
@@ -97,64 +94,28 @@ namespace cuda {
 // -------------------------------------------------------------
 #ifdef __CUDACC__
 
-// Atomic add functions
-#if __CUDA_ARCH__ >= 530
-__device__ __inline__ __half atomic_add(__half* address, __half val) {
-#if 0 // TODO: replace this once Nvidia implements atomicAdd for __half
-  return atomicAdd(address, val);
-#else
-  unsigned int* address_as_uint = (unsigned int*) address;
-  unsigned int old = *address_as_uint;
-  __half* old_as_half = (__half*) &old;
-  unsigned int assumed;
-  unsigned int updated;
-  __half* updated_as_half = (__half*) &updated;
-  do {
-    assumed = old;
-    updated = old;
-    *updated_as_half += val;
-    old = atomicCAS(address_as_uint, assumed, updated);
-  } while (assumed != old);
-  return *old_as_half;
-#endif // 0
-}
-#endif // __CUDA_ARCH__ >= 530
-__device__ __inline__ float atomic_add(float* address, float val) {
-  return atomicAdd(address, val);
-}
-__device__ __inline__ double atomic_add(double* address, double val) {
-#if __CUDA_ARCH__ >= 600
-  return atomicAdd(address, val);
-#else
-  unsigned long long int* address_as_ull =
-    (unsigned long long int*)address;
-  unsigned long long int old = *address_as_ull, assumed;
-  do {
-    assumed = old;
-    old = atomicCAS(address_as_ull, assumed,
-                    __double_as_longlong(val +
-                                         __longlong_as_double(assumed)));
-  } while (assumed != old);
-  return __longlong_as_double(old);
-#endif // __CUDA_ARCH__ < 600
-}
+// Atomic add
+template <typename T> __device__ __inline__
+T atomic_add(T* address, T val);
 
 // Min and max
 template <typename T> __device__ __inline__
 T min(const T& x, const T& y) { return x <= y ? x : y; }
 template <typename T> __device__ __inline__
 T max(const T& x, const T& y) { return x >= y ? x : y; }
-template <> __device__ __inline__
-float min<float>(const float& x, const float& y) { return fminf(x, y); }
-template <> __device__ __inline__
-double min<double>(const double& x, const double& y) { return fmin(x, y); }
-template <> __device__ __inline__
-float max<float>(const float& x, const float& y) { return fmaxf(x, y); }
-template <> __device__ __inline__
-double max<double>(const double& x, const double& y) { return fmax(x, y); }
+
+// Numeric limits
+template <typename T> constexpr __device__ __inline__ T min();
+template <typename T> constexpr __device__ __inline__ T max();
+template <typename T> constexpr __device__ __inline__ T epsilon();
+template <typename T> __device__ __inline__ T infinity();
   
 #endif // __CUDACC__
 
+// -------------------------------------------------------------
+// Utilities for CUDA events
+// -------------------------------------------------------------
+  
 /** Wrapper class for a CUDA event. */
 class event_wrapper {
 public:
@@ -186,96 +147,13 @@ private:
 // -------------------------------------------------------------
 #ifdef __CUDACC__
 
-/** CUDA kernel to apply an entry-wise unary operator. */
-template <typename UnaryOperator>
-__global__
-void entrywise_unary_operator_kernel(El::Int height, El::Int width,
-                                     const DataType* __restrict__ input,
-                                     El::Int input_ldim,
-                                     DataType* __restrict__ output,
-                                     El::Int output_ldim) {
-  const El::Int gid = threadIdx.x + blockIdx.x * blockDim.x;
-  const El::Int size = height * width;
-  const El::Int num_threads = blockDim.x * gridDim.x;
-  UnaryOperator op;
-  for (El::Int pos = gid; pos < size; pos += num_threads) {
-    const auto& row = pos % height;
-    const auto& col = pos / height;
-    const auto& x = input[row + col * input_ldim];
-    auto& y = output[row + col * output_ldim];
-    y = op(x);
-  }
-}
-
-/** CUDA kernel to apply an entry-wise binary operator. */
-template <typename BinaryOperator>
-__global__
-void entrywise_binary_operator_kernel(El::Int height, El::Int width,
-                                     const DataType* __restrict__ input1,
-                                     El::Int input1_ldim,
-                                     const DataType* __restrict__ input2,
-                                     El::Int input2_ldim,
-                                     DataType* __restrict__ output,
-                                     El::Int output_ldim) {
-  const El::Int gid = threadIdx.x + blockIdx.x * blockDim.x;
-  const El::Int size = height * width;
-  const El::Int num_threads = blockDim.x * gridDim.x;
-  BinaryOperator op;
-  for (El::Int pos = gid; pos < size; pos += num_threads) {
-    const auto& row = pos % height;
-    const auto& col = pos / height;
-    const auto& x1 = input1[row + col * input1_ldim];
-    const auto& x2 = input2[row + col * input2_ldim];
-    auto& y = output[row + col * output_ldim];
-    y = op(x1, x2);
-  }
-}
-
 /** Apply an entry-wise unary operator to GPU data.
  *  The input and output data must be on GPU and must have the same
  *  dimensions.
  */
 template <typename UnaryOperator>
 void apply_entrywise_unary_operator(const AbsMat& input,
-                                    AbsMat& output) {
-
-  // Check that input and output are valid
-  std::stringstream err;
-  if (input.GetDevice() != El::Device::GPU) {
-    LBANN_ERROR("input is not on GPU");
-  } else if (output.GetDevice() != El::Device::GPU) {
-    LBANN_ERROR("output is not on GPU");
-  } else if (input.Height() != output.Height()
-             || input.Width() != output.Width()) {
-    err << "input matrix dimensions "
-        << "(" << input.Height() << " x " << input.Width() << ")"
-        << "don't match output matrix dimensions "
-        << "(" << output.Height() << " x " << output.Width() << ")";
-    LBANN_ERROR(err.str());
-  }
-
-  // Get CUDA grid dimensions
-  // Note: Maximum CUDA grid dimension is 2^32-1
-  // (https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#features-and-technical-specifications).
-  const El::Int height = input.Height();
-  const El::Int width = input.Width();
-  const El::Int block_dim = 256;
-  El::Int grid_dim = (height * width + block_dim - 1) / block_dim;
-  if (sizeof(El::Int) > sizeof(unsigned int)
-      && grid_dim > std::numeric_limits<uint32_t>::max()) {
-    grid_dim = std::numeric_limits<uint32_t>::max();
-  }
-
-  // Launch CUDA kernel
-  if (grid_dim > 0) {
-    CHECK_CUDA(cudaSetDevice(El::GPUManager::Device()));
-    entrywise_unary_operator_kernel<UnaryOperator>
-      <<<grid_dim, block_dim, 0, El::GPUManager::Stream()>>>(
-        height, width, input.LockedBuffer(), input.LDim(),
-        output.Buffer(), output.LDim());
-  }
-  
-}
+                                    AbsMat& output);
 
 /** Apply an entry-wise binary operator to GPU data.
  *  The input and output data must be on GPU and must have the same
@@ -284,74 +162,16 @@ void apply_entrywise_unary_operator(const AbsMat& input,
 template <typename BinaryOperator>
 void apply_entrywise_binary_operator(const AbsMat& input1,
                                      const AbsMat& input2,
-                                     AbsMat& output) {
-
-  // Check that input and output are valid
-  std::stringstream err;
-  if (input1.GetDevice() != El::Device::GPU
-      || input2.GetDevice() != El::Device::GPU) {
-    LBANN_ERROR("input is not on GPU");
-  } else if (output.GetDevice() != El::Device::GPU) {
-    LBANN_ERROR("output is not on GPU");
-  } else if (input1.Height() != input2.Height()
-             || input1.Width() != input2.Width()
-             || input1.Height() != output.Height()
-             || input1.Width() != output.Width()) {
-    err << "input matrix dimensions "
-        << "(" << input1.Height() << " x " << input1.Width() << ", "
-        << input2.Height() << " x " << input2.Width() << ")"
-        << "don't match output matrix dimensions "
-        << "(" << output.Height() << " x " << output.Width() << ")";
-    LBANN_ERROR(err.str());
-  }
-
-  // Get CUDA grid dimensions
-  // Note: Maximum CUDA grid dimension is 2^32-1
-  // (https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#features-and-technical-specifications).
-  const El::Int height = input1.Height();
-  const El::Int width = input1.Width();
-  const El::Int block_dim = 256;
-  El::Int grid_dim = (height * width + block_dim - 1) / block_dim;
-  if (sizeof(El::Int) > sizeof(unsigned int)
-      && grid_dim > std::numeric_limits<uint32_t>::max()) {
-    grid_dim = std::numeric_limits<uint32_t>::max();
-  }
-
-  // Launch CUDA kernel
-  if (grid_dim > 0) {
-    CHECK_CUDA(cudaSetDevice(El::GPUManager::Device()));
-    entrywise_binary_operator_kernel<BinaryOperator>
-      <<<grid_dim, block_dim, 0, El::GPUManager::Stream()>>>(
-        height, width,
-        input1.LockedBuffer(), input1.LDim(),
-        input2.LockedBuffer(), input2.LDim(),
-        output.Buffer(), output.LDim());
-  }
+                                     AbsMat& output);
   
-}
 
-  
 /** Apply an entry-wise unary operator to GPU data.
  *  The input and output data must be on GPU, have the same
  *  dimensions, and be aligned.
  */
 template <typename UnaryOperator>
 void apply_entrywise_unary_operator(const AbsDistMat& input,
-                                    AbsDistMat& output) {
-  std::stringstream err;
-  if (input.Height() != output.Height()
-      || input.Width() != output.Width()) {
-    err << "input matrix dimensions "
-        << "(" << input.Height() << " x " << input.Width() << ")"
-        << "don't match output matrix dimensions "
-        << "(" << output.Height() << " x " << output.Width() << ")";
-    LBANN_ERROR(err.str());
-  } else if (input.DistData() != output.DistData()) {
-    LBANN_ERROR("input and output matrix distributions don't match");
-  }
-  apply_entrywise_unary_operator<UnaryOperator>(input.LockedMatrix(),
-                                                output.Matrix());
-}
+                                    AbsDistMat& output);
 
 /** Apply an entry-wise binary operator to GPU data.
  *  The input and output data must be on GPU, have the same
@@ -360,26 +180,7 @@ void apply_entrywise_unary_operator(const AbsDistMat& input,
 template <typename BinaryOperator>
 void apply_entrywise_binary_operator(const AbsDistMat& input1,
                                      const AbsDistMat& input2,
-                                     AbsDistMat& output) {
-  if (input1.Height() != input2.Height()
-      || input1.Width() != input2.Width()
-      || input1.Height() != output.Height()
-      || input1.Width() != output.Width()) {
-    std::stringstream err;
-    err << "input matrix dimensions "
-        << "(" << input1.Height() << " x " << input1.Width() << ", "
-        << input2.Height() << " x " << input2.Width() << ")"
-        << "don't match output matrix dimensions "
-        << "(" << output.Height() << " x " << output.Width() << ")";
-    LBANN_ERROR(err.str());
-  } else if (input1.DistData() != input2.DistData()
-             || input1.DistData() != output.DistData()) {
-    LBANN_ERROR("input and output matrix distributions don't match");
-  }
-  apply_entrywise_binary_operator<BinaryOperator>(input1.LockedMatrix(),
-                                                  input2.LockedMatrix(),
-                                                  output.Matrix());
-}
+                                     AbsDistMat& output);
   
 #endif // __CUDACC__
 
@@ -415,35 +216,12 @@ public:
     : m_stream(stream) {}
 
   /** Allocate GPU buffer. */
-  pointer allocate(size_type size) {
-    value_type* buffer = nullptr;
-    if (size > 0) {
-#ifdef HYDROGEN_HAVE_CUB
-      auto& memory_pool = El::cub::MemoryPool();
-      CHECK_CUDA(memory_pool.DeviceAllocate(reinterpret_cast<void**>(&buffer),
-                                            size * sizeof(value_type),
-                                            m_stream));
-#else
-      CHECK_CUDA(cudaMalloc(&buffer, size * sizeof(value_type)));
-#endif // HYDROGEN_HAVE_CUB
-    }
-    return pointer(buffer);
-  }
+  pointer allocate(size_type size);
 
   /** Deallocate GPU buffer.
    *  'size' is unused and maintained for compatibility with Thrust.
    */
-  void deallocate(pointer buffer, size_type size = 0) {
-    auto&& ptr = buffer.get();
-    if (ptr != nullptr) {
-#ifdef HYDROGEN_HAVE_CUB
-      auto& memory_pool = El::cub::MemoryPool();
-      CHECK_CUDA(memory_pool.DeviceFree(ptr));
-#else
-      CHECK_CUDA(cudaFree(ptr));
-#endif // HYDROGEN_HAVE_CUB
-    }
-  }
+  void deallocate(pointer buffer, size_type size = 0);
 
 };
 
@@ -451,6 +229,9 @@ public:
 
 } // namespace cuda
 } // namespace lbann
+
+// Template implementations
+#include "lbann/utils/impl/cuda.hpp"
 
 #endif // LBANN_HAS_GPU
 #endif // LBANN_UTILS_CUDA_HPP
